refact(SinglePokemonCard): created card component for Single Pokemon Page
